ordo-backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LangChain and LangGraph for AI orchestration
langchain==0.1.4
langchain-core==0.1.16
langchain-mistralai==0.0.5
langgraph==0.0.20
langchain-mcp-adapters==0.1.0
langsmith==0.0.77

# Mistral AI
mistralai==0.1.3

# HTTP client
httpx==0.26.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Supabase for RAG
supabase==2.3.4
pgvector==0.2.4

# Redis for caching
redis==5.0.1
hiredis==2.3.2

# Security and validation
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Rate limiting
slowapi==0.1.9

# CORS and middleware
python-cors==1.0.0

# Monitoring and logging
python-json-logger==4.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
hypothesis==6.98.3

# MCP Server support
fastmcp==0.1.0
mcp==0.1.0

# Solana
solders==0.20.0
solana==0.32.0

# Utilities
python-dateutil==2.8.2

# Gmail API Integration
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# X/Twitter API Integration
tweepy==4.14.0

# Telegram Bot API
python-telegram-bot==20.7
